requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-google-genai
chromadb
sentence-transformers
fastapi
uvicorn
python-dotenv
pandas
